# docs/swagger.yaml
definitions:
...
  models.Response:
    properties:
      body: {}
      message:
        type: string
      status:
        type: boolean
    type: object
  models.ServiceCreate:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  models.ServiceLaundry:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.ServiceUpdate:
    properties:
      id:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
info:
  contact: {}
  description: This is a api to app gestioncar
  termsOfService: http://swagger.io/terms/
  title: APP GESTIONCAR
  version: "1.0"
paths:
...
  /service/{id}:
    get:
      consumes:
      - application/json
      description: Get a service by id
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      - description: ID of the income to get
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                body:
                  $ref: '#/definitions/models.ServiceLaundry'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Service not found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get a service by id
      tags:
      - Service
  /service/create:
    post:
      consumes:
      - application/json
      description: Creates a service and returns its ID.
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      - description: Service creation data
        in: body
        name: serviceCreate
        required: true
        schema:
          $ref: '#/definitions/models.ServiceCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Service created successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                body:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Model is invalid
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Create Service
      tags:
      - Service
  /service/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a service based on the provided ID and workplace context.
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      - description: ID of the service
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Servicio eliminado con éxito
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Service not found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Error interno
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Delete Service
      tags:
      - Service
  /service/get_all:
    get:
      consumes:
      - application/json
      description: Get all services from workplace
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/models.ServiceLaundry'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get all services from workplace
      tags:
      - Service
  /service/update:
    put:
      consumes:
      - application/json
      description: Updates the details of a service based on the provided data.
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      - description: Service data to update
        in: body
        name: serviceUpdate
        required: true
        schema:
          $ref: '#/definitions/models.ServiceUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Servicio editado con éxito
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Expense not found
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Model is invalid
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Error interno
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Update Service
      tags:
      - Service
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT token. Example: "Bearer
      eyJhbGciOiJIUz..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"